Separate PATCH utils into submodule of v1::utils
