+++ server/svix-server/src/v1/endpoints/application.rs
@@ -14,9 +14,12 @@ use crate::{
     db::models::application,
     error::{HttpError, Result},
     v1::utils::{
-        patch_field_non_nullable, patch_field_nullable, validate_no_control_characters,
-        validate_no_control_characters_unrequired, EmptyResponse, ListResponse, ModelIn, ModelOut,
-        Pagination, PaginationLimit, UnrequiredField, UnrequiredNullableField, ValidatedJson,
+        patch::{
+            patch_field_non_nullable, patch_field_nullable, UnrequiredField,
+            UnrequiredNullableField,
+        },
+        validate_no_control_characters, validate_no_control_characters_unrequired, EmptyResponse,
+        ListResponse, ModelIn, ModelOut, Pagination, PaginationLimit, ValidatedJson,
         ValidatedQuery,
     },
 };
